move roi manager
